+++ src/main/java/de/mth/game/gameobject/Bullet.java
@@ -7,11 +7,11 @@ import de.mth.game.texture.TextureLoader;
 
 public class Bullet extends GameObject {
 
-	public Bullet(int x, int y, float destX, float destY) {
+	public Bullet(int x, int y, double destX, double destY) {
 		super(x, y);
 		setSpeed(3);
 
-		float[] vel = calculateVelocity(destX, destY);
+		double[] vel = calculateVelocity(destX, destY);
 
 		setVelX(vel[0]);
 		setVelY(vel[1]);
